backend/core/requirements.txt
boto3>=1.26.43
uvicorn>=0.20.0
SQLAlchemy>=1.4.38
sqlalchemy-pytds>=0.3.4
python-tds>=1.11.0
pyopenssl>=22.1.0
pytz>=2022.7
SQLAlchemy-Utils>=0.36.8
psycopg2-binary>=2.8.5
aiofiles>=0.5.0
certifi>=2020.4.5.1
chardet>=3.0.4
Click>=7.0
colorama>=0.4.3
h11>=0.9.0
httptools>=0.0.13
idna>=2.8
prompt-toolkit>=2.0.10
Pygments>=2.3.1
python-dateutil>=2.8.1
python-multipart>=0.0.5
pytz>=2020.1
PyYAML>=5.3.1
requests>=2.22.0
six>=1.15.0
starlette>=0.13.2
urllib3>=1.24.3
uvicorn>=0.11.2
uvloop>=0.14.0
wcwidth>=0.1.9
websockets>=8.1
PyJWT>=2.1.0
passlib>=1.7.4
bcrypt>=3.2.0
pypng>=0.20220715.0
pytz>=2023.3
merkle-json>=1.0.0
fhir.resources==7.0.2
PyQRCode==1.2.1
cryptography==41.0.4
pillow==10.2.0
fastapi==0.109.1
pydantic==1.10.14
PyPDF2==3.0.1
jellyfish==1.0.3
fuzzywuzzy==0.18.0
openai>=1.13.3